refactor(upgrade): use dynamic import for subcommand
