+++ packages/opencode/src/index.ts
@@ -6,7 +6,7 @@ import * as Log from "@opencode-ai/core/util/log"
 import { ConsoleCommand } from "./cli/cmd/account/command"
 import { ProvidersCommand } from "./cli/cmd/providers/command"
 import { AgentCommand } from "./cli/cmd/agent/command"
-import { UpgradeCommand } from "./cli/cmd/upgrade"
+import { UpgradeCommand } from "./cli/cmd/upgrade/command"
 import { UninstallCommand } from "./cli/cmd/uninstall/command"
 import { ModelsCommand } from "./cli/cmd/models/command"
 import { UI } from "./cli/ui"
